1-kafka-admin-fast-api/requirements.txt
fastapi[all]
aiokafka
kafka-python
python-dotenv
